Add curves and itemsquisheras to items JSON

+++ apps/tool/Models/Curves/DumpCurvePoint.cs
@@ -11,14 +11,14 @@ public class DumpCurvePoint
     public int OrderIndex { get; set; }
 
     [Name("Pos[0]")]
-    public int Pos0 { get; set; }
+    public double Pos0 { get; set; }
 
     [Name("Pos[1]")]
-    public int Pos1 { get; set; }
+    public double Pos1 { get; set; }
 
     [Name("PosPreSquish[0]")]
-    public int PosPreSquish0 { get; set; }
+    public double PosPreSquish0 { get; set; }
 
     [Name("PosPreSquish[1]")]
-    public int PosPreSquish1 { get; set; }
+    public double PosPreSquish1 { get; set; }
 }
